Move deprecated JSON-comparing ShouldMatch to the single place it is used, to limit further usage.
